control-port-filter-python.service.d systemd drop-in file
Workaround:
control-port-filter-python fails if it cannot bind on the network interface.
This is the case when networking did not come up, which is the case when
whonix-firewall.service failed. To make control-port-filter-python
automatically start after lets says Whonix firewall config was fixed and the
whonix-firewall.service was restarted.

After=whonix-firewall.service
Requires=whonix-firewall.service